requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-phonenumber-field